+++ TerraeRasa/TerraeRasa/src/render/RenderUI.java
@@ -331,7 +331,7 @@ public class RenderUI extends Render
 			if(chest.metaData != 1) //Make sure its metadata is 1 (otherwise it doesnt technically exist)
 			{
 				//Get the metadata for the block's size
-				int[][] metadata = MetaDataHelper.getMetaDataArray(world.getBlock(player.viewedChestX, player.viewedChestY).blockWidth / 6, world.getBlock(player.viewedChestX, player.viewedChestY).blockHeight / 6); //metadata used by the block of size (x,y)
+				int[][] metadata = MetaDataHelper.getMetaDataArray((int)(world.getBlock(player.viewedChestX, player.viewedChestY).blockWidth / 6), (int)(world.getBlock(player.viewedChestX, player.viewedChestY).blockHeight / 6)); //metadata used by the block of size (x,y)
 				int metaWidth = metadata.length; 
 				int metaHeight = metadata[0].length;	
 				x1 = 0;
@@ -1965,7 +1965,7 @@ public class RenderUI extends Render
 		if(chest.metaData != 1) //Make sure its metadata is 1 (otherwise it doesnt technically exist)
 		{
 			//Get the metadata for the block's size
-			int[][] metadata = MetaDataHelper.getMetaDataArray(world.getBlock(player.viewedChestX, player.viewedChestY).blockWidth / 6, world.getBlock(player.viewedChestX, player.viewedChestY).blockHeight / 6); //metadata used by the block of size (x,y)
+			int[][] metadata = MetaDataHelper.getMetaDataArray((int)(world.getBlock(player.viewedChestX, player.viewedChestY).blockWidth / 6), (int)(world.getBlock(player.viewedChestX, player.viewedChestY).blockHeight / 6)); //metadata used by the block of size (x,y)
 			int metaWidth = metadata.length; 
 			int metaHeight = metadata[0].length;	
 			int x1 = 0;
@@ -2029,7 +2029,7 @@ public class RenderUI extends Render
 		if(chest.metaData != 1)
 		{
 			//Get the metadata for the block's size
-			int[][] metadata = MetaDataHelper.getMetaDataArray(world.getBlock(player.viewedChestX, player.viewedChestY).blockWidth / 6, world.getBlock(player.viewedChestX, player.viewedChestY).blockHeight / 6); //metadata used by the block of size (x,y)
+			int[][] metadata = MetaDataHelper.getMetaDataArray((int)(world.getBlock(player.viewedChestX, player.viewedChestY).blockWidth / 6), (int)(world.getBlock(player.viewedChestX, player.viewedChestY).blockHeight / 6)); //metadata used by the block of size (x,y)
 			int metaWidth = metadata.length; 
 			int metaHeight = metadata[0].length;	
 			int x = 0;
